requirements.txt
websockets
psycopg2-binary
aiohttp